Update Power of Four.c

+++ C#/Power of Four.c
@@ -5,7 +5,7 @@ bool isPowerOfFour(int n) {
         }
     for(int i=0;i<=16;i++)
         {
-        if(pow(4,i)==n)
+        if((int)pow(4,i)==n)
             {
             return 1;
             }
